refactor: Improved quick sort for readability

+++ sort/quick_sort/quick.h
@@ -2,6 +2,6 @@
 #define QUICK_H
 
 // Quick Sort prototypes
-void quick_sort(int arr[], size_t length);
+void quick_sort(int arr[], int length);
 
 #endif
